Initial commit: Add all files necessary to generate PDF
This generates a PDF file of wedding invitations. It reads the data from
a CSV file invite-list which contains columns for name, language and
lagzi. It uses the tex template files to build one big LaTeX file which
can then be compiled to PDF.

The language and lagzi columns determine which templates will be used.
In Hungary, the lagzi is a big party after the wedding ceremonies and
not everybody who attends is the ceremonies is necessarily invited to
the lagzi. The language columns doesn't really need to be just a
language. In this example I use en (for English) and hu (for Hungarian)
but in reality we'll use hu_s_inf, hu_pl_inf, hu_s_fml, hu_s_fml because
Hungarian invitations will need different text for formal, informal,
singular and plural, whereas the English ones don't. It can be anything,
the filename just needs to match. The lagzi invitation text is read from
a file language_lagzi.tex where langauge is the language for this person
and _lagzi is the actual text '_lagzi'.

The PDF is set up to print out two invitations per page in landscape
mode with a cursive text and with space in the middle to fold it in
half. It has space for an address, invitation text, a poem or verse and
a link to a homepage for more details